services/gradio_sql/requirements.txt
gradio
psycopg2-binary
dash
dotenv
sqlalchemy